requirements.txt
django==1.4.0
django-reversion==1.6.4
git+git://github.com/luminousflux/overdiff.git#egg=overdiff
south
git+git://github.com/luminousflux/django-taggit.git#egg=django-taggit
markdown
-e git+git://github.com/fin/django-pagedown.git#egg=django-pagedown
PIL
-e git+git://github.com/luminousflux/django-userena.git#egg=django-userena
django-admin-tools
django-debug-toolbar
django-nose
-e git+git://github.com/luminousflux/django-luminous-tumblelog.git#egg=django-luminous-tumblelog
embedly
django-voting
-e git+git://github.com/luminousflux/django-threadedcomments.git#egg=django-threadedcomments
#tastypie HEAD( >.9.11)  has SessionAuthentication
-e git+git://github.com/toastdriven/django-tastypie.git#egg=django-tastypie
tldextract
# installed but not required. only injects code if configured in settings:
django-googleanalytics
django-overextends
django-celery
django-celery-email
html2text
